wx v2.7.0 incremental update
git-svn-id: http://svn.osafoundation.org/chandler/trunk@9781 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ internal/wx/src/motif/radiobox.cpp
@@ -4,7 +4,7 @@
 // Author:      Julian Smart
 // Modified by:
 // Created:     17/09/98
-// RCS-ID:      $Id: radiobox.cpp,v 1.50 2006/03/14 19:44:35 ABX Exp $
+// RCS-ID:      $Id: radiobox.cpp,v 1.51 2006/03/15 00:01:18 ABX Exp $
 // Copyright:   (c) Julian Smart
 // Licence:     wxWindows licence
 /////////////////////////////////////////////////////////////////////////////
@@ -198,7 +198,7 @@ void wxRadioBox::SetSelection(int n)
     XmToggleButtonSetState ((Widget) m_radioButtons[n], True, False);
 
     for (size_t i = 0; i < m_noItems; i++)
-        if (i != n)
+        if (i != (size_t)n)
             XmToggleButtonSetState ((Widget) m_radioButtons[i], False, False);
 
     m_inSetValue = false;
@@ -381,11 +381,11 @@ void wxRadioBoxCallback (Widget w, XtPointer clientData,
 
   wxRadioBox *item = (wxRadioBox *) clientData;
   int sel = -1;
-  int i;
+  size_t i;
   const wxWidgetArray& buttons = item->GetRadioButtons();
   for (i = 0; i < item->GetCount(); i++)
     if (((Widget)buttons[i]) == w)
-      sel = i;
+      sel = (int)i;
   item->SetSel(sel);
 
   if (item->InSetValue())
